+++ sample/src/commonMain/kotlin/com/github/panpf/sketch/sample/ui/test/AnimatedImageTestScreen.kt
@@ -156,7 +156,6 @@ class AnimatedImageTestScreen : BaseScreen() {
                                     components {
                                         add(gifDecoder)
                                     }
-                                    repeatCount(0)
                                     onAnimationStart(startCallback)
                                     onAnimationEnd(endCallback)
                                 },