+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Translations: [简体中文](CHANGELOG.zh.md)
      prompt for upgrade.
 > 3. Reference [《Migration Documentation》](docs/migrate.md) migrating from 3.x to 4.x
 
+# new
+
+* fix: Fixed the bug that Koral GifDrawable would not actively call back onAnimationStart and
+  onAnimationEnd during automatic playback. [#278](https://github.com/panpf/sketch/issues/278)
+
 # 4.4.0-beta01
 
 There are no changes compared to version 4.4.0-alpha02